requirements.txt
fastapi
uvicorn
captcha
asyncpg
pyyaml
python-multipart